// node/common/src/client/models/SessionSpec.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Theia.cloud API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.8.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists } from '../runtime';
/**
 * 
 * @export
 * @interface SessionSpec
 */
export interface SessionSpec {
    /**
     * 
     * @type {string}
     * @memberof SessionSpec
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof SessionSpec
     */
    appDefinition?: string;
    /**
     * 
     * @type {string}
     * @memberof SessionSpec
     */
    user?: string;
    /**
     * 
     * @type {string}
     * @memberof SessionSpec
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof SessionSpec
     */
    error?: string;
    /**
     * 
     * @type {string}
     * @memberof SessionSpec
     */
    workspace?: string;
    /**
     * 
     * @type {number}
     * @memberof SessionSpec
     */
    lastActivity?: number;
}

/**
 * Check if a given object implements the SessionSpec interface.
 */
export function instanceOfSessionSpec(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function SessionSpecFromJSON(json: any): SessionSpec {
    return SessionSpecFromJSONTyped(json, false);
}

export function SessionSpecFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionSpec {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': !exists(json, 'name') ? undefined : json['name'],
        'appDefinition': !exists(json, 'appDefinition') ? undefined : json['appDefinition'],
        'user': !exists(json, 'user') ? undefined : json['user'],
        'url': !exists(json, 'url') ? undefined : json['url'],
        'error': !exists(json, 'error') ? undefined : json['error'],
        'workspace': !exists(json, 'workspace') ? undefined : json['workspace'],
        'lastActivity': !exists(json, 'lastActivity') ? undefined : json['lastActivity'],
    };
}

export function SessionSpecToJSON(value?: SessionSpec | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'appDefinition': value.appDefinition,
        'user': value.user,
        'url': value.url,
        'error': value.error,
        'workspace': value.workspace,
        'lastActivity': value.lastActivity,
    };
}