// node/common/src/client/models/SessionStartRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Theia.cloud API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.8.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists } from '../runtime';
/**
 * A request to start a session
 * @export
 * @interface SessionStartRequest
 */
export interface SessionStartRequest {
    /**
     * The App Id of this Theia.cloud instance. Request without a matching Id will be denied.
     * @type {string}
     * @memberof SessionStartRequest
     */
    appId: string;
    /**
     * The user identification, usually the email address.
     * @type {string}
     * @memberof SessionStartRequest
     */
    user: string;
    /**
     * The app to launch.
     * @type {string}
     * @memberof SessionStartRequest
     */
    appDefinition: string;
    /**
     * The name of the workspace to mount/create.
     * @type {string}
     * @memberof SessionStartRequest
     */
    workspaceName?: string;
    /**
     * Number of minutes to wait for session launch. Default is 3 Minutes.
     * @type {number}
     * @memberof SessionStartRequest
     */
    timeout?: number;
}

/**
 * Check if a given object implements the SessionStartRequest interface.
 */
export function instanceOfSessionStartRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "appId" in value;
    isInstance = isInstance && "user" in value;
    isInstance = isInstance && "appDefinition" in value;

    return isInstance;
}

export function SessionStartRequestFromJSON(json: any): SessionStartRequest {
    return SessionStartRequestFromJSONTyped(json, false);
}

export function SessionStartRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionStartRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'appId': json['appId'],
        'user': json['user'],
        'appDefinition': json['appDefinition'],
        'workspaceName': !exists(json, 'workspaceName') ? undefined : json['workspaceName'],
        'timeout': !exists(json, 'timeout') ? undefined : json['timeout'],
    };
}

export function SessionStartRequestToJSON(value?: SessionStartRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'appId': value.appId,
        'user': value.user,
        'appDefinition': value.appDefinition,
        'workspaceName': value.workspaceName,
        'timeout': value.timeout,
    };
}